AD-288 refactor errors
